requirements.txt
aiofiles
httpx
pyuseragents
bs4
loguru
tortoise-orm
yaml